requirements.txt
setuptools==65.6.3
requests==2.32.3
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-system-metrics
opentelemetry-instrumentation-grpc
